+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/Implementation/RedisProfilerEntryToActivityConverter.cs
@@ -101,14 +101,17 @@ internal static class RedisProfilerEntryToActivityConverter
             StackExchangeRedisConnectionInstrumentation.ActivitySourceNew :
             StackExchangeRedisConnectionInstrumentation.ActivitySource;
 
+        var creationTags =
+            options.EmitOldAttributes && options.EmitNewAttributes ? StackExchangeRedisConnectionInstrumentation.BothCreationTags :
+            options.EmitNewAttributes ? StackExchangeRedisConnectionInstrumentation.NewCreationTags :
+            options.EmitOldAttributes ? StackExchangeRedisConnectionInstrumentation.OldCreationTags :
+            [];
+
         var activity = activitySource.StartActivity(
             name,
             ActivityKind.Client,
             parentActivity?.Context ?? default,
-            [
-                .. options.EmitOldAttributes ? StackExchangeRedisConnectionInstrumentation.OldCreationTags : [],
-                .. options.EmitNewAttributes ? StackExchangeRedisConnectionInstrumentation.NewCreationTags : [],
-            ],
+            creationTags,
             startTime: command.CommandCreated);
 
         if (activity == null)
@@ -120,11 +123,11 @@ internal static class RedisProfilerEntryToActivityConverter
 
         if (activity.IsAllDataRequested)
         {
-            // see https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/database.md
+            // See https://github.com/open-telemetry/semantic-conventions/blob/v1.42.0/docs/db/database-spans.md
 
             // Timing example:
-            // command.CommandCreated; //2019-01-10 22:18:28Z
 
+            // command.CommandCreated;          // 2019-01-10 22:18:28Z
             // command.CreationToEnqueued;      // 00:00:32.4571995
             // command.EnqueuedToSending;       // 00:00:00.0352838
             // command.SentToResponse;          // 00:00:00.0060586
@@ -169,7 +172,7 @@ internal static class RedisProfilerEntryToActivityConverter
 
             if (options.EmitNewAttributes)
             {
-                string? queryText = command.Command;
+                var queryText = command.Command;
                 if (options.SetVerboseDatabaseStatements && !string.IsNullOrEmpty(commandAndKey))
                 {
                     queryText = commandAndKey;
@@ -189,9 +192,10 @@ internal static class RedisProfilerEntryToActivityConverter
             {
                 if (command.EndPoint is IPEndPoint ipEndPoint)
                 {
-                    activity.SetTag(SemanticConventions.AttributeServerAddress, ipEndPoint.Address.ToString());
+                    var address = ipEndPoint.Address.ToString();
+                    activity.SetTag(SemanticConventions.AttributeServerAddress, address);
                     activity.SetTag(SemanticConventions.AttributeServerPort, ipEndPoint.Port);
-                    activity.SetTag(SemanticConventions.AttributeNetworkPeerAddress, ipEndPoint.Address.ToString());
+                    activity.SetTag(SemanticConventions.AttributeNetworkPeerAddress, address);
                     activity.SetTag(SemanticConventions.AttributeNetworkPeerPort, ipEndPoint.Port);
                 }
                 else if (command.EndPoint is DnsEndPoint dnsEndPoint)
@@ -202,8 +206,9 @@ internal static class RedisProfilerEntryToActivityConverter
 #if NET
                 else if (command.EndPoint is UnixDomainSocketEndPoint unixDomainSocketEndPoint)
                 {
-                    activity.SetTag(SemanticConventions.AttributeServerAddress, unixDomainSocketEndPoint.ToString());
-                    activity.SetTag(SemanticConventions.AttributeNetworkPeerAddress, unixDomainSocketEndPoint.ToString());
+                    var address = unixDomainSocketEndPoint.ToString();
+                    activity.SetTag(SemanticConventions.AttributeServerAddress, address);
+                    activity.SetTag(SemanticConventions.AttributeNetworkPeerAddress, address);
                 }
 #endif
             }