+++ src/OpenTelemetry.Sampler.AWS/CHANGELOG.md
@@ -9,6 +9,9 @@
   missing or null `Attributes`.
   ([#4316](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4316))
 
+* Updated OpenTelemetry core component version(s) to `1.16.0`.
+  ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+
 ## 0.1.0-alpha.9
 
 Released 2026-Apr-21